requirements.txt
wxPython
tendo
flask
pywebview
pyinstaller
pandas
pyarrow
openpyxl